refactor(core): Reduce code duplication in DB config (no-changelog) (#8679)
