--- packages/cli/src/databases/ormconfig.ts
@@ -1,13 +1,0 @@
-import {
-	getMariaDBConnectionOptions,
-	getMysqlConnectionOptions,
-	getPostgresConnectionOptions,
-	getSqliteConnectionOptions,
-} from './config';
-
-export default [
-	getSqliteConnectionOptions(),
-	getPostgresConnectionOptions(),
-	getMysqlConnectionOptions(),
-	getMariaDBConnectionOptions(),
-];
